Split out internal classes into their own class (for CatalogBundleTracker and CatalogBundleLoader)
